+++ CLibs/linked_list/linked_list.h
@@ -1,8 +1,6 @@
 #ifndef _LL_ONE_WAY_HEADER_H_
 #define _LL_ONE_WAY_HEADER_H_
 
-#include <stdlib.h>
-#include <errno.h>
 #include <stdbool.h>
 
 typedef struct int_linked_list_node_s {
@@ -54,7 +52,7 @@ bool ll_erase_value(int_linked_list_t *ll, int value_to_erase);
 /// @brief Erase value by index in linked list
 /// @param ll Object into which erasing will be performed
 /// @param index Index at which erasure will be performed
-void ll_erase_value_by_idx(int_linked_list_t *ll, int index);
+void ll_erase_value_by_idx(int_linked_list_t *ll, size_t index);
 
 /// @brief Prints all items of linked list with new-line symbol at the end
 /// @param ll Object in which embedded payload will be output
