+++ CLibs/linked_list_two_way/linked_list_2.c
@@ -157,7 +157,7 @@ bool ll2_erase_value(int_linked_list2_t *ll, int value_to_erase) {
     return true;
 }
 
-void ll2_erase_value_by_idx(int_linked_list2_t *ll, int index) {
+void ll2_erase_value_by_idx(int_linked_list2_t *ll, size_t index) {
     if (index >= ll->size) {
         return;
     }
@@ -199,7 +199,7 @@ void ll2_print(int_linked_list2_t *ll) {
 
     int_linked_list_node2_t *iterator = ll->sentinel_node->next_node;
     printf("[ ");
-    for (int i = 0; i < ll->size-1; i++) {
+    for (size_t i = 0; i < ll->size-1; i++) {
         printf("%d <-> ", iterator->payload);
         iterator = iterator->next_node;
     }
